Refactor converters into partial classes
