+++ OfficeIMO.Word.Html/OfficeIMO.Word.Html.csproj
@@ -39,6 +39,13 @@
     </ItemGroup>
 
     <ItemGroup>
+        <Compile Update="Converters/HtmlToWordConverter.Tables.cs" />
+        <Compile Update="Converters/HtmlToWordConverter.Lists.cs" />
+        <Compile Update="Converters/HtmlToWordConverter.Images.cs" />
+        <Compile Update="Converters/HtmlToWordConverter.Formatting.cs" />
+    </ItemGroup>
+
+    <ItemGroup>
         <PackageReference Include="DocumentFormat.OpenXml" Version="[3.3.0,4.0.0)" />
         <PackageReference Include="AngleSharp" Version="1.3.0" />
     </ItemGroup>
