+++ OfficeIMO.Word.Markdown/Converters/MarkdownToWordConverter.cs
@@ -1,12 +1,9 @@
 using Markdig;
+using Markdig.Extensions.Tables;
 using Markdig.Syntax;
 using Markdig.Syntax.Inlines;
 using OfficeIMO.Word;
 using System;
-using System.IO;
-using System.Linq;
-using System.Text;
-using Markdig.Extensions.Tables;
 
 namespace OfficeIMO.Word.Markdown.Converters {
     /// <summary>
@@ -23,7 +20,7 @@ namespace OfficeIMO.Word.Markdown.Converters {
     ///    - LinkInline -> paragraph.AddHyperLink()
     /// 4. Reuse existing OfficeIMO.Word functionality, don't recreate
     /// </summary>
-    internal class MarkdownToWordConverter {
+    internal partial class MarkdownToWordConverter {
         public WordDocument Convert(string markdown, MarkdownToWordOptions options) {
             if (markdown == null) {
                 throw new ArgumentNullException(nameof(markdown));
@@ -60,17 +57,7 @@ namespace OfficeIMO.Word.Markdown.Converters {
                     ProcessInline(paragraphBlock.Inline, listItemParagraph, options, document);
                     break;
                 case ListBlock listBlock:
-                    var list = listBlock.IsOrdered ? document.AddListNumbered() : document.AddListBulleted();
-                    foreach (ListItemBlock listItem in listBlock) {
-                        var firstParagraph = listItem.FirstOrDefault() as ParagraphBlock;
-                        if (firstParagraph != null) {
-                            var listParagraph = list.AddItem(string.Empty, listLevel);
-                            ProcessInline(firstParagraph.Inline, listParagraph, options, document);
-                        }
-                        foreach (var sub in listItem.Skip(1)) {
-                            ProcessBlock(sub, document, options, list, listLevel + 1);
-                        }
-                    }
+                    ProcessListBlock(listBlock, document, options, listLevel);
                     break;
                 case QuoteBlock quote:
                     foreach (var sub in quote) {
@@ -97,104 +84,5 @@ namespace OfficeIMO.Word.Markdown.Converters {
                     break;
             }
         }
-
-        private static void ProcessTable(Table table, WordDocument document, MarkdownToWordOptions options) {
-            int rows = table.Count();
-            int cols = table.ColumnDefinitions.Count;
-            var wordTable = document.AddTable(rows, cols);
-            int r = 0;
-            foreach (TableRow row in table) {
-                int c = 0;
-                foreach (TableCell cell in row) {
-                    var target = wordTable.Rows[r].Cells[c].Paragraphs[0];
-                    foreach (var cellBlock in cell) {
-                        if (cellBlock is ParagraphBlock pb) {
-                            ProcessInline(pb.Inline, target, options, document);
-                        }
-                    }
-                    c++;
-                }
-                r++;
-            }
-        }
-
-        private static void ProcessInline(Inline? inline, WordParagraph paragraph, MarkdownToWordOptions options, WordDocument document) {
-            if (inline == null) {
-                return;
-            }
-
-            var buffer = new StringBuilder();
-
-            void Flush() {
-                if (buffer.Length > 0) {
-                    InlineRunHelper.AddInlineRuns(paragraph, buffer.ToString(), options.FontFamily);
-                    buffer.Clear();
-                }
-            }
-
-            for (var current = inline; current != null; current = current.NextSibling) {
-                if (current is LinkInline link) {
-                    Flush();
-                    if (link.IsImage) {
-                        AddImage(document, paragraph, link);
-                    } else {
-                        string label = BuildMarkdown(link.FirstChild);
-                        var hyperlink = paragraph.AddHyperLink(label, new Uri(link.Url, UriKind.RelativeOrAbsolute));
-                        if (!string.IsNullOrEmpty(options.FontFamily)) {
-                            hyperlink.SetFontFamily(options.FontFamily);
-                        }
-                    }
-                } else {
-                    buffer.Append(BuildMarkdown(current));
-                }
-            }
-            Flush();
-        }
-
-        private static void AddImage(WordDocument document, WordParagraph paragraph, LinkInline link) {
-            if (File.Exists(link.Url)) {
-                paragraph.AddImage(link.Url);
-            } else {
-                document.AddImageFromUrl(link.Url, 50, 50);
-            }
-        }
-
-        private static string BuildMarkdown(Inline? inline) {
-            if (inline == null) {
-                return string.Empty;
-            }
-
-            var sb = new StringBuilder();
-            for (var current = inline; current != null; current = current.NextSibling) {
-                switch (current) {
-                    case LiteralInline literal:
-                        sb.Append(literal.Content.ToString());
-                        break;
-                    case EmphasisInline emphasis:
-                        string marker = new('*', emphasis.DelimiterCount);
-                        sb.Append(marker);
-                        sb.Append(BuildMarkdown(emphasis.FirstChild));
-                        sb.Append(marker);
-                        break;
-                    case LineBreakInline:
-                        sb.Append('\n');
-                        break;
-                    case ContainerInline container:
-                        sb.Append(BuildMarkdown(container.FirstChild));
-                        break;
-                }
-            }
-
-            return sb.ToString();
-        }
-
-        private static string GetCodeBlockText(CodeBlock codeBlock) {
-            var sb = new StringBuilder();
-            foreach (var line in codeBlock.Lines.Lines) {
-                sb.AppendLine(line.Slice.ToString());
-            }
-            return sb.ToString().TrimEnd();
-        }
-
     }
 }
